requirements.txt
numpy>=1.26
openai>=1.40.0
pinecone>=5.0.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
sentence-transformers>=3.0.1
